merging

+++ parsing/GNL/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: fstitou <fstitou@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/03 22:11:15 by fstitou           #+#    #+#             */
-/*   Updated: 2022/11/01 22:44:17 by fstitou          ###   ########.fr       */
+/*   Updated: 2022/11/04 04:26:17 by fstitou          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ char	*get_next_line(int fd);
 char	*ft_strjoin(char const *s, char const *s1);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strdup(const char *s);
-size_t	ft_strlen(const char *s);
+int		ft_strlen(const char *s);
 int		ft_strchr(const char *s, int c);
 
 #endif
